+++ etc/statsd_fake.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require "socket"
+
+Socket.udp_server_loop(8045) do |data, src|
+  puts data
+end