+++ metrics/prometheus.go
@@ -15,6 +15,8 @@ const (
 func (m *Metrics) Prometheus() string {
 	var buf strings.Builder
 
+	tags := toPromTags(m.tags)
+
 	m.EachCounter(func(counter *Counter) {
 		name := prometheusNamespace + `_` + counter.Name()
 
@@ -22,7 +24,7 @@ func (m *Metrics) Prometheus() string {
 			"\n# HELP " + name + " " + counter.Desc() + "\n",
 		)
 		buf.WriteString("# TYPE " + name + " counter\n")
-		buf.WriteString(name + " " + strconv.FormatUint(counter.Value(), 10) + "\n")
+		buf.WriteString(name + tags + " " + strconv.FormatUint(counter.Value(), 10) + "\n")
 	})
 
 	m.EachGauge(func(gauge *Gauge) {
@@ -32,7 +34,7 @@ func (m *Metrics) Prometheus() string {
 			"\n# HELP " + name + " " + gauge.Desc() + "\n",
 		)
 		buf.WriteString("# TYPE " + name + " gauge\n")
-		buf.WriteString(name + " " + strconv.FormatUint(gauge.Value(), 10) + "\n")
+		buf.WriteString(name + tags + " " + strconv.FormatUint(gauge.Value(), 10) + "\n")
 	})
 
 	return buf.String()
@@ -43,4 +45,20 @@ func (m *Metrics) PrometheusHandler(w http.ResponseWriter, r *http.Request) {
 	metricsData := m.Prometheus()
 
 	fmt.Fprint(w, metricsData)
+}
+
+func toPromTags(tags map[string]string) string {
+	if tags == nil {
+		return ""
+	}
+
+	buf := make([]string, len(tags))
+	i := 0
+
+	for k, v := range tags {
+		buf[i] = fmt.Sprintf("%s=\"%s\"", k, v)
+		i++
+	}
+
+	return fmt.Sprintf("{%s}", strings.Join(buf, ", "))
 }