+++ metrics/prometheus_test.go
@@ -3,9 +3,12 @@ package metrics
 import (
 	"net/http"
 	"net/http/httptest"
+	"regexp"
+	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestPrometheus(t *testing.T) {
@@ -52,6 +55,34 @@ anycable_go_tests 123
 anycable_go_any_tests 0
 `,
 	)
+}
+
+func TestPrometheusWithTags(t *testing.T) {
+	m := NewMetrics(nil, 10)
+	m.DefaultTags(map[string]string{"env": "dev", "instance": "R2D2"})
+
+	m.RegisterCounter("test_total", "Total number of smth")
+	m.RegisterGauge("tests", "Number of active smth")
+
+	m.Gauge("tests").Set(123)
+	m.Counter("test_total").Add(3)
+
+	actual := m.Prometheus()
+
+	r := regexp.MustCompile(`anycable_go_test_total{(.+)}\s+(\d+)`)
+	matches := r.FindStringSubmatch(actual)
+
+	require.NotNil(t, matches)
+
+	tagsStr := matches[1]
+	valStr := matches[2]
+
+	tags := strings.Split(tagsStr, ", ")
+
+	assert.Contains(t, tags, `env="dev"`)
+	assert.Contains(t, tags, `instance="R2D2"`)
+
+	assert.Equal(t, "3", valStr)
 }
 
 func TestPrometheusHandler(t *testing.T) {