+++ metrics/statsd_writer.go
@@ -1,6 +1,7 @@
 package metrics
 
 import (
+	"fmt"
 	"strings"
 	"sync"
 
@@ -11,6 +12,7 @@ import (
 type StatsdConfig struct {
 	Host          string
 	Prefix        string
+	TagFormat     string
 	MaxPacketSize int
 }
 
@@ -29,7 +31,7 @@ func (lg *StatsdLogger) Printf(msg string, args ...interface{}) {
 }
 
 func NewStatsdConfig() StatsdConfig {
-	return StatsdConfig{Prefix: "anycable_go.", MaxPacketSize: 1400}
+	return StatsdConfig{Prefix: "anycable_go.", MaxPacketSize: 1400, TagFormat: "datadog"}
 }
 
 func (c StatsdConfig) Enabled() bool {
@@ -39,26 +41,48 @@ func (c StatsdConfig) Enabled() bool {
 type StatsdWriter struct {
 	client *statsd.Client
 	config StatsdConfig
+	tags   map[string]string
 
 	mu sync.Mutex
 }
 
 var _ IntervalWriter = (*StatsdWriter)(nil)
 
-func NewStatsdWriter(c StatsdConfig) *StatsdWriter {
-	return &StatsdWriter{config: c}
+func NewStatsdWriter(c StatsdConfig, tags map[string]string) *StatsdWriter {
+	return &StatsdWriter{config: c, tags: tags}
 }
 
 func (sw *StatsdWriter) Run(interval int) error {
 	sl := StatsdLogger{log.WithField("context", "statsd")}
-	sw.client = statsd.NewClient(
-		sw.config.Host,
+	opts := []statsd.Option{
 		statsd.MaxPacketSize(sw.config.MaxPacketSize),
 		statsd.MetricPrefix(sw.config.Prefix),
 		statsd.Logger(&sl),
+	}
+
+	var tagsInfo string
+
+	if sw.tags != nil {
+		tagsStyle := resolveTagsStyle(sw.config.TagFormat)
+		tags := convertTags(sw.tags)
+		opts = append(opts,
+			statsd.TagStyle(tagsStyle),
+			statsd.DefaultTags(tags...),
+		)
+
+		tagsInfo = fmt.Sprintf(", tags=%v, style=%s", sw.tags, sw.config.TagFormat)
+	}
+
+	sw.client = statsd.NewClient(
+		sw.config.Host,
+		opts...,
 	)
 
-	log.WithField("context", "metrics").Infof("Send statsd metrics to %s with prefix %s every %vs", sw.config.Host, sw.config.Prefix, interval)
+	log.WithField("context", "metrics").
+		Infof(
+			"Send statsd metrics to %s with every %vs (prefix=%s%s)",
+			sw.config.Host, interval, sw.config.Prefix, tagsInfo,
+		)
 
 	return nil
 }
@@ -88,4 +112,29 @@ func (sw *StatsdWriter) Write(m *Metrics) error {
 	})
 
 	return nil
+}
+
+func resolveTagsStyle(name string) *statsd.TagFormat {
+	switch name {
+	case "datadog":
+		return statsd.TagFormatDatadog
+	case "influxdb":
+		return statsd.TagFormatInfluxDB
+	case "graphite":
+		return statsd.TagFormatGraphite
+	}
+
+	panic(fmt.Errorf("Unknown StatsD tags format: %s", name))
+}
+
+func convertTags(tags map[string]string) []statsd.Tag {
+	buf := make([]statsd.Tag, len(tags))
+	i := 0
+
+	for k, v := range tags {
+		buf[i] = statsd.StringTag(k, v)
+		i++
+	}
+
+	return buf
 }