+++ metrics/statsd_writer_test.go
@@ -26,7 +26,7 @@ func TestStatsdWriter(t *testing.T) {
 	t.Run("Write send UDP with metrics", func(t *testing.T) {
 		c := NewStatsdConfig()
 		c.Host = socket.LocalAddr().String()
-		w := NewStatsdWriter(c)
+		w := NewStatsdWriter(c, nil)
 		_ = w.Run(0)
 		defer w.Stop()
 
@@ -52,7 +52,7 @@ func TestStatsdWriter(t *testing.T) {
 		c := NewStatsdConfig()
 		c.Host = socket.LocalAddr().String()
 		c.Prefix = "ws."
-		w := NewStatsdWriter(c)
+		w := NewStatsdWriter(c, nil)
 		_ = w.Run(0)
 		defer w.Stop()
 
@@ -72,6 +72,112 @@ func TestStatsdWriter(t *testing.T) {
 
 		assert.Contains(t, payload, "ws.test_count:10|c")
 		assert.Contains(t, payload, "ws.test_gauge:123|g")
+	})
+
+	t.Run("Adds tags in DataDog format", func(t *testing.T) {
+		c := NewStatsdConfig()
+		c.TagFormat = "datadog"
+		c.Host = socket.LocalAddr().String()
+		w := NewStatsdWriter(c, map[string]string{"env": "dev"})
+		_ = w.Run(0)
+		defer w.Stop()
+
+		err := w.Write(m)
+		assert.NoError(t, err)
+
+		var buf []byte
+
+		select {
+		case buf = <-received:
+		case <-time.After(time.Second):
+			t.Error("timeout waiting for UDP payload")
+			return
+		}
+
+		payload := string(buf)
+
+		assert.Contains(t, payload, "anycable_go.test_count:10|c|#env:dev")
+		assert.Contains(t, payload, "anycable_go.test_gauge:123|g|#env:dev")
+	})
+
+	t.Run("Adds multiple tags in DataDog format", func(t *testing.T) {
+		c := NewStatsdConfig()
+		c.TagFormat = "datadog"
+		c.Host = socket.LocalAddr().String()
+		w := NewStatsdWriter(c, map[string]string{"env": "dev", "rev": "1.1"})
+		_ = w.Run(0)
+		defer w.Stop()
+
+		err := w.Write(m)
+		assert.NoError(t, err)
+
+		var buf []byte
+
+		select {
+		case buf = <-received:
+		case <-time.After(time.Second):
+			t.Error("timeout waiting for UDP payload")
+			return
+		}
+
+		payload := string(buf)
+
+		assert.Contains(t, payload, "anycable_go.test_count:10|c|#")
+		assert.Contains(t, payload, "anycable_go.test_gauge:123|g|#")
+		assert.Contains(t, payload, "env:dev")
+		assert.Contains(t, payload, "rev:1.1")
+	})
+
+	t.Run("Adds tags in InfluxDB format", func(t *testing.T) {
+		c := NewStatsdConfig()
+		c.TagFormat = "influxdb"
+		c.Host = socket.LocalAddr().String()
+		w := NewStatsdWriter(c, map[string]string{"env": "dev"})
+		_ = w.Run(0)
+		defer w.Stop()
+
+		err := w.Write(m)
+		assert.NoError(t, err)
+
+		var buf []byte
+
+		select {
+		case buf = <-received:
+		case <-time.After(time.Second):
+			t.Error("timeout waiting for UDP payload")
+			return
+		}
+
+		payload := string(buf)
+
+		assert.Contains(t, payload, "anycable_go.test_count,env=dev:10|c")
+		assert.Contains(t, payload, "anycable_go.test_gauge,env=dev:123|g")
+	})
+
+	t.Run("Adds tags in Graphite format", func(t *testing.T) {
+		c := NewStatsdConfig()
+		c.TagFormat = "graphite"
+		c.Host = socket.LocalAddr().String()
+		w := NewStatsdWriter(c, map[string]string{"env": "dev"})
+		_ = w.Run(0)
+		defer w.Stop()
+
+		err := w.Write(m)
+		assert.NoError(t, err)
+
+		var buf []byte
+
+		select {
+		case buf = <-received:
+		case <-time.After(time.Second):
+			t.Error("timeout waiting for UDP payload")
+			return
+		}
+
+		payload := string(buf)
+
+		assert.Contains(t, payload, "anycable_go.test_count;env=dev:10|c")
+		assert.Contains(t, payload, "anycable_go.test_gauge;env=dev:123|g")
 	})
 }